AppDelegate main QuizController property
Note this wouldn’t be like this if we wanted to have multiple Quizs in
the same game. In that case I would recommend having a Quizs controller
which returns the name of all our Quizs and show them in the first view
controller within a table view. Once the user selects one of them, we
would initialise it with its name.
But since we just have a single Quiz for the whole app, it may be
initialised and shared from AppDelegate